added reflections, refractions, cube primitive and loading scene from json

+++ Engine/src/Texture/UVImage.cpp
@@ -12,5 +12,5 @@ Color UVImage::atUV(const UV & uv) const
     double v = 1 - uv.v;
     double x = uv.u * (image->width - 1);
     double y = v * (image->height - 1);
-    return image->at(round(x), round(y));
+    return image->at(static_cast<int>(round(x)), static_cast<int>(round(y)));
 }
